group common parsers by theme
