+++ common/characters.test.ts
@@ -1,7 +1,20 @@
 import { assertEquals } from "@std/assert";
-import { lower } from "./mod.ts";
 import { parseErrors } from "../errors.ts";
 import { twoLower } from "../tests/common.test.ts";
+import { letter, lower, upper } from "./mod.ts";
+
+Deno.test("letter", () => {
+  assertEquals(letter.parse("m"), {
+    success: true,
+    results: [{ value: "m", remaining: "", position: { line: 1, column: 1 } }],
+  });
+
+  assertEquals(letter.parse("1"), {
+    success: false,
+    message: parseErrors.letter,
+    position: { line: 1, column: 0 },
+  });
+});
 
 Deno.test("lower", () => {
   assertEquals(lower.parse("Hello"), {
@@ -23,5 +36,17 @@ Deno.test("lower", () => {
     success: false,
     message: "Expected two lowercase letters",
     position: { line: 1, column: 0 },
+  });
+});
+
+
+Deno.test("upper", () => {
+  assertEquals(upper.parse("Hello"), {
+    success: true,
+    results: [{
+      value: "H",
+      remaining: "ello",
+      position: { line: 1, column: 1 },
+    }],
   });
 });
