--- common/decimal.test.ts
@@ -1,29 +1,0 @@
-import { assertEquals } from "@std/assert/equals";
-import { decimal } from "./mod.ts";
-import { parseErrors } from "../errors.ts";
-
-Deno.test("decimal", () => {
-  assertEquals(decimal.parse("2.3 and more"), {
-    success: true,
-    results: [{
-      value: 2.3,
-      remaining: "and more",
-      position: { line: 1, column: 4 },
-    }],
-  });
-
-  assertEquals(decimal.parse("23 and more"), {
-    success: false,
-    message: parseErrors.decimal,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(decimal.parse("-2.3 and more"), {
-    success: true,
-    results: [{
-      value: -2.3,
-      remaining: "and more",
-      position: { line: 1, column: 5 },
-    }],
-  });
-});
