--- common/spaces.ts
@@ -1,8 +1,0 @@
-import { regex } from "./mod.ts";
-import type { Parser } from "../index.ts";
-
-/**
- * Parses the space character (0 or more)
- */
-
-export const spaces: Parser<string> = regex(/^ */);
